Add copied logic
A previous attempt was made at this project. The code
has been copied from that test repo. The other repo failed
because I had organized things in such a way that it
was necessary to pass objects from a child component
to a parent component. I only found out that is not
possible in React after it was too late to be salvaged.